requirements.txt
azure-core==1.6.0
azure-cosmos==4.0.0
azure-functions==1.2.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
dateparser==0.7.6
google-api-core==1.21.0
google-api-python-client==1.7.9
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.18.1
humanize==2.5.0
idna==2.9
oauth2client==4.1.3
oauthlib==3.1.0
pendulum==2.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
pytzdata==2020.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sendgrid==6.3.1
six==1.15.0
snaptime==0.2.4
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
